requirements.txt
pybullet==3.2.5
jupyter==1.0.0
jupyterlab==3.4.4
matplotlib==3.5.2
stable-baselines3==1.6.0
torch==1.12.0
numpy==1.22.0
pandas==1.4.3
Pillow==9.0.0
ipython==8.4.0
ipykernel==6.15.1
gym==0.21.0
nbconvert==6.5.0
notebook==6.4.12
pip==21.2.4
PyYAML==6.0
pyzmq==24.0.1
